# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Catalog API",
    "description": "Microservicio de Catalog - ECommerce Architecture - Soporte multiidioma con Accept-Language header (es, en)",
    "version": "v1"
  },
  "paths": {
    "/v1/categories": {
      "get": {
        "tags": [
          "Category"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDtoDataCollection"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDtoDataCollection"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDtoDataCollection"
                }
              }
            }
          }
        }
      }
    },
    "/v1/categories/{id}": {
      "get": {
        "tags": [
          "Category"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/Home": {
      "get": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "es"
            }
          },
          {
            "name": "productsPerSection",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 8
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HomePageResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomePageResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomePageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v1/Home/banners": {
      "get": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "position",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "hero"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "es"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BannerDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BannerDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BannerDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/Home/featured": {
      "get": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 8
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "es"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/Home/deals": {
      "get": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 8
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "es"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/Home/bestsellers": {
      "get": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 8
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "es"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/Home/new-arrivals": {
      "get": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 8
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "es"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/Home/top-rated": {
      "get": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 8
            }
          },
          {
            "name": "minRating",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 4
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "es"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/Home/categories": {
      "get": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 8
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "es"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BannerDto": {
        "type": "object",
        "properties": {
          "bannerId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "subtitle": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "imageUrlMobile": {
            "type": "string",
            "nullable": true
          },
          "linkUrl": {
            "type": "string",
            "nullable": true
          },
          "buttonText": {
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CategoryDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "parentCategoryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "isFeatured": {
            "type": "boolean"
          },
          "productCount": {
            "type": "integer",
            "format": "int32"
          },
          "subCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            },
            "nullable": true
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "breadcrumbs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryBreadcrumbDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CategoryDtoDataCollection": {
        "type": "object",
        "properties": {
          "hasItems": {
            "type": "boolean",
            "readOnly": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            },
            "nullable": true
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "HomePageResponse": {
        "type": "object",
        "properties": {
          "banners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BannerDto"
            },
            "nullable": true
          },
          "featuredProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "nullable": true
          },
          "deals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "nullable": true
          },
          "bestSellers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "nullable": true
          },
          "newArrivals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "nullable": true
          },
          "topRated": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "nullable": true
          },
          "featuredCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            },
            "nullable": true
          },
          "metadata": {
            "$ref": "#/components/schemas/HomeMetadata"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Ingrese el token JWT en el formato: Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "ApiKey": {
        "type": "apiKey",
        "description": "Ingrese el API Key para comunicación entre servicios",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ],
      "ApiKey": [ ]
    }
  ]
}